Update jdownloader.sh
Did not need the second argument of "hidden" after all, as it implies the end users input will be hidden.
Switched to swizzin default method of apt_update and apt_install. Which should cut down on the output to the cli.
Send output of the initial JDownloader run to the swizzin log rather than the cli.
$user was the wrong variable to grab the username from swizzin. $1 is the correct variable.
A couple other minor adjustments.