installs jdownloader for each user now
Added importing swizzin's "user" functions as well as "globals.sh". Not sure if it's necessary or not, as some functions seemed to work without importing them.
Wrapped the JDownloader installation into a function, and made it install for each user when the script is called from 'box install jdownloader'.
Should also now install for a new user, when a new user is added via 'box adduser'.