changed install location from 'jd' to 'jd2'
This is just so that the install location will align with where JDownloader's own shell script would install it. I figure this might help in case their is a tutorial in their docs or forums that might assume someone has it installed in the default place.
Added an additional reference.