rTorrent: Disable warning messages during compile
Warning messages require additional code checks and are echoed to the log file. This requires additional time and resources to compile. If we want warning messages, we could create a developer mode for swizzin. These have no place for production purposes, as they do not impact functionality of the software.