rTorrent: Installation speed improvements (#962)
* rTorrent: Compile speed improvement 

Compile libtorrent & rtorrent with level 1 optimizations if there are 2 or less processor threads available.

This drastically reduces build times on systems such as virtual machines. It is still better than level 0 repo.

* rTorrent: Compile with pipe improvements

* rTorrent: Disable warning messages during compile

Warning messages require additional code checks and are echoed to the log file. This requires additional time and resources to compile. If we want warning messages, we could create a developer mode for swizzin. These have no place for production purposes, as they do not impact functionality of the software.

* rTorrent: Retreive using curl & silence extraction

We can retrieve the files faster using curl instead of wget. This is already installed as a rtorrent dependency.

The extraction process is also faster if we don't verbose what we're doing. The log file is cleaner as well.

* Reduce pipe requirement to 512MB

* rTorrent: Use curl to retreive mktorrent source

* Fix typo at command end

* Add configure_rtorrent function

* Fixes for configure_rtorrent