rTorrent: Install faster with ramdisk
This commit mounts a 1 GB tmpfs ram disk for rtorrent installations.  Four gigabytes of system memory is required. Increases installation speed and efficiency.

It was required to refactor the build paths. The ram disk needs to be mounted in a non conflicting location.

Also, it creates a new `rtorrent.log` file for the rtorrent installation. The log file is written to the ram disk, then copied to the logs folder afterwards. This resolves the megabyte log file problem with multiple installs and reduces disk writes.