rTorrent: Add UDNS install option
This commit adds an option to install a stability patch for UDP trackers on rTorrent. It's back ported against version 0.9.8.

With this patch, the user will be able to distribute thousands of torrents with UDP trackers, without running into software stability issues.

That being said, there are still two unsolved race conditions. This patch is not at fault for causing the crashes.

Install advisory is updated to inform users about the option. libudns built from source for compatibility with Debian. It takes a few seconds and is only built if the option is enabled.