ruTorrent: Install ffmpeg by default
ffmpeg is required for the screenshots plugin. Many users want to use this plugin, so it's not a good idea to gut the functionality. It's supported on all platforms and is a tiny binary file.

This commit makes the ffmpeg install script modular, so it's easier to maintain and changes apply to ruTorrent. It only installs ffmpeg if it's not already installed.