rTorrent: Upgrade GCC if supported
This script upgrades GCC on Ubuntu 20.04 or 22.04 LTS using packages. It improves the performance of software compiled with GCC. This task is performed relatively quickly.

Manual upgrade option is added. It also does it automatically if rTorrent is being compiled from source.

Ubuntu 20.04 LTS seamlessly goes from GCC 9 to 10. Ubuntu 22.04 LTS seamlessly goes from GCC 11 to 12.

update-alternatives are configured to allow smooth transition from Ubuntu 20.04 to 22.04 LTS. The user just needs to run the script again for the newer version to take priority over the old.

Debian has chosen not to provide packages for upgrading GCC. Therefore, it's not possible to seamlessly support this OS. It would take hours to manually compile GCC.