ruTorrent: Set $localHostedMode to true (#984)
It was introduced and set to false when testing ruTorrent version 4.0.1. The regressions were since resolved in ruTorrent version 4.0.2.
See change-log here: https://github.com/Novik/ruTorrent/releases/tag/v4.0.2

The purpose of this to improve loading times when ruTorrent is hosted on the SAME machine as rTorrent. It performs plugin dependency checks through shell and causes the results. This is a more viable option than executing XMLRPC commands, which is designed to only be performed for remote hosts.