rTorrent: Add UDNS install option (#982)
* rTorrent: Add UDNS install option

This commit adds an option to install a stability patch for UDP trackers on rTorrent. It's back ported against version 0.9.8.

With this patch, the user will be able to distribute thousands of torrents with UDP trackers, without running into software stability issues.

That being said, there are still two unsolved race conditions. This patch is not at fault for causing the crashes.

Install advisory is updated to inform users about the option. libudns built from source for compatibility with Debian. It takes a few seconds and is only built if the option is enabled.

* Fix formatting issue with sh-checker

* rTorrent: Fix libtorrent scanf crash

This commit fixes a race condition with scanf inside libtorrent. It uses a vector to cache the udp tracker info to reduce the probably of a crash by approximately 98%.

* rTorrent: Disable forced tracker startup scrape

This patch avoids a thread timeout crash at startup by delaying tracker updates. They will be queued into normal intervals, instead of being forcefully updated at startup.

It also allows instantaneous startup times in situations where it would normally take minutes to process thousands of torrents.