rTorrent: Consolidate package installs
This commit consolidates package installs to be performed at the same time. It reduces install script overhead and makes it easier for the user to understand what is happening.

Feature specific requirements are defined with a new variable, so developers know which ones to adjust if maintenance is required.

Compatible with #1003.