rTorrent: Network configuration
This commit implements a 4MB send and receive buffer for TCP sockets. It improves throughput on rTorrent. Debian/Ubuntu will automatically downscale this value if the user doesn't have enough memory. The maximum default without changes is 4MB.

It also adds a missing configuration variable for per torrent upload slots. The default value of 50 is applied.